- Wrote test for checking bug: https://github.com/healenium/healenium-web/issues/30 Problem with hover - Some tests moved in new class ElementsAndActions - Added methods for MainPageWithFindBy - Fixed testFindElementWithinParent
