+++ build.gradle
@@ -18,6 +18,6 @@ repositories {
 dependencies {
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.hamcrest', name: 'hamcrest-core', version: '2.1'
-    compile 'com.epam.healenium:healenium-web:2.0.4.1'
+    compile 'com.epam.healenium:healenium-web:2.0.5'
     testCompile 'com.codeborne:selenide:5.10.0'
 }
