+++ src/main/java/pages/MainPageWithFindBy.java
@@ -3,6 +3,7 @@ package pages;
 
 import com.epam.healenium.SelfHealingDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -43,6 +44,16 @@ public class MainPageWithFindBy extends BasePage {
         } catch (Exception e) {
             return false;
         }
+    }
+
+    public MainPageWithFindBy hoverOnTestButton() {
+        Actions action = new Actions(driver);
+        action.moveToElement(testButton).perform();
+        return this;
+    }
+
+    public String getAttributeFromTestButton(String attribute) {
+        return testButton.getAttribute(attribute);
     }
 
 }
