+++ src/test/resources/healenium.properties
@@ -1,4 +1,5 @@
 recovery-tries = 3
+score-cap = .5
 basePath = build/sha/selenium
 reportPath = build/reports
 screenshotPath = build/screenshots/
